refactor: create HomeViewModel and refactor HomeView
