+++ StayFit360/StayFit360/Home/Models/Activity.swift
@@ -1,0 +1,17 @@
+//
+//  Activity.swift
+//  StayFit360
+//
+//  Created by Pratik Haridas Ghadge on 03/10/25.
+//
+
+import SwiftUI
+
+struct Activity {
+	let id: Int
+	let title: String
+	let subtitle: String
+	let image: String
+	let tintColor: Color
+	let amount: String
+}
