+++ StayFit360/StayFit360/Home/Models/Workout.swift
@@ -1,0 +1,18 @@
+//
+//  Workout.swift
+//  StayFit360
+//
+//  Created by Pratik Haridas Ghadge on 03/10/25.
+//
+
+import SwiftUI
+
+struct Workout {
+	let id: Int
+	let title: String
+	let image: String
+	let tintColor: Color
+	let duration: String
+	let date: String
+	let calories: String
+}
